setup uses ts
